pyproject.toml
[project]
name = "enzona-py-sdk"
version = "0.1.0"
description = "Python SDK for the Enzona Cuban payment API"
requires-python = ">=3.8"
dependencies = [
    "httpx>=0.23,<0.25",
    "pydantic>=1.10,<2",
    "tenacity>=8.0,<9",
    "python-dotenv>=0.19,<2",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0,<8",
    "pytest-asyncio>=0.21,<1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["enzona_service"]
